+++ src/content/config.ts
@@ -1,7 +1,44 @@
-import { markketplace } from "@config";
 import { defineCollection } from "astro:content";
+import { markketplace } from "@config";
+import {
+  fetchStrapiContent,
+  fetchStrapiSchema,
+  type StrapiQueryOptions,
+  type StrapiConfig
+} from 'cafecito';
 
-import { strapiLoader } from "../lib/strapi-loader";
+const config: StrapiConfig = {
+  store_slug: markketplace.STORE_SLUG,
+  api_url: markketplace.STRAPI_URL,
+  sync_interval: 6000,
+};
+
+/**
+ * Creates a Strapi content loader for Astro
+ * @param contentType The Strapi content type to load
+ * @param filter The filter to apply to the content &filters[store][id][$eq]=${STRAPI_STORE_ID}
+ * @returns An Astro loader for the specified content type
+ */
+function strapiLoader(query: StrapiQueryOptions) {
+  return {
+    name: `strapi-${query.contentType}`,
+    schema: async () => await fetchStrapiSchema(query.contentType, config.api_url),
+
+    async load({ store, logger, meta }) {
+      const lastSynced = meta.get("lastSynced");
+      if (lastSynced && Date.now() - Number(lastSynced) < config.sync_interval) {
+        logger.info("Skipping sync");
+        return;
+      }
+
+      const posts = await fetchStrapiContent(query, config);
+
+      store.clear();
+      posts.forEach((item: any) => store.set({ id: item.id, data: item }));
+      meta.set("lastSynced", String(Date.now()));
+    },
+  };
+};
 
 const pages = defineCollection({
   loader: strapiLoader({
